Fix a bug in conversion types for float32 classes

+++ src/main/java/nom/bdezonia/zorbage/type/data/float32/octonion/OctonionFloat32MatrixMember.java
@@ -282,7 +282,7 @@ public final class OctonionFloat32MatrixMember
 	
 	@Override
 	public PrimitiveRepresentation preferredRepresentation() {
-		return PrimitiveRepresentation.DOUBLE;
+		return PrimitiveRepresentation.FLOAT;
 	}
 
 	@Override
