+++ src/main/java/nom/bdezonia/zorbage/type/data/float32/octonion/OctonionFloat32RModuleMember.java
@@ -256,7 +256,7 @@ public final class OctonionFloat32RModuleMember
 	
 	@Override
 	public PrimitiveRepresentation preferredRepresentation() {
-		return PrimitiveRepresentation.DOUBLE;
+		return PrimitiveRepresentation.FLOAT;
 	}
 
 	@Override
