+++ src/main/java/nom/bdezonia/zorbage/type/data/float32/quaternion/QuaternionFloat32Member.java
@@ -219,7 +219,7 @@ public final class QuaternionFloat32Member
 
 	@Override
 	public PrimitiveRepresentation preferredRepresentation() {
-		return PrimitiveRepresentation.DOUBLE;
+		return PrimitiveRepresentation.FLOAT;
 	}
 
 	@Override
