+++ src/main/java/nom/bdezonia/zorbage/type/data/float32/quaternion/QuaternionFloat32RModuleMember.java
@@ -240,7 +240,7 @@ public final class QuaternionFloat32RModuleMember
 	
 	@Override
 	public PrimitiveRepresentation preferredRepresentation() {
-		return PrimitiveRepresentation.DOUBLE;
+		return PrimitiveRepresentation.FLOAT;
 	}
 
 	@Override
